return native queries information from the schema endpoint (#6)
### What

The [schema
endpoint](https://hasura.github.io/ndc-spec/specification/schema/index.html)
expects information about all collections in the `scalar_types`,
`object_types`, and `collections` fields. Previously, we only returned
the information about tables. In this PR we include the information
about the native queries as well.

We also properly return the nullability of a column or argument in this
PR.

### How

The schema endpoints expects three things that are relevant for us atm:

- scalar_types - we need to fetch the scalar types that are visible in
native queries (from columns or arguments) and include them.
- object_types - define the "return type" of the collection (name and
fields), so we need to return the native queries and their columns.
- collections - defines the "return type" of the collection (the
matching object type), as well as the arguments to the native query.