Start services in parallel, not serially.
In testing, we start each database and Jaeger one by one, waiting for
the health checks for each one. This is slooooow.

If we just combine the commands, it's way faster.

We use Docker Compose as intended, specifying all the required services
on one line.

I have also copied the Jaeger service definition out of the v3-engine
file so we can just use it directly.