Add `order_by` clauses absolutely everywhere. (#22)
### What

We can't have deterministic snapshots without `ORDER BY`. PostgreSQL is
generous and generally orders by the primary key for small datasets, but
Yugabyte and other databases are not so forgiving.

### How

I have added `order_by` clauses to the goldenfiles pretty much wherever
we expect multiple results.

The `EXPLAIN` snapshots needed to be updated, but none others.