Run tests against a Yugabyte container. (#23)
### What

Yugabyte fully supports PostgreSQL, so no changes are needed to the
connector itself, but we do need to ensure that everything works as
expected.

This adds test cases for Yugabyte and the associated snapshots (which
are identical to the PostgreSQL snapshots).

### How

I copied the test definitions from the AWS Aurora definitions, and the
snapshots from the PostgreSQL connector.

The tests are packaged inside the `other-db-tests` crate, similarly to
the AWS Aurora tests. I've added a feature, `yugabyte`, so you need to
opt in. `just test` will do this automatically if you're running on an
x86_64 processor, as it doesn't seem to work on macOS + aarch64 through
emulation.

There's a lot of copying and pasting here. This is probably a good
example of where we might want to combine snapshots, and potentially
test case definitions.