CI: Build the tests once. (#28)
### What

`cargo nextest` has the ability to build a test archive, containing all
the compiled binaries, and then use it on a subsequent run. We can use
this to avoid having to recompile tests for each job, which speeds up
the test runs considerably and avoids re-work.

### How

We introduce a new job which builds and archives the tests. Subsequent
jobs download the archive and run the tests, using filters to select the
packages.

We need to switch to filters because the `-p` flag is not supported with
archives.

The archive, built in release mode, weighs in at around 160 MB, which is
sizeable, but still reasonable. (In debug mode, it's over 1 GB, which is
_not_ reasonable.)