Skip deserializing and re-serializing JSON query responses.
We have introduced a new feature in ndc-sdk which wraps responses in a
`JsonResponse`. This allows us to either hand a serializable value to
`axum` to serialize into JSON and respond to the client, as before, or
optionally pass through bytes which we assume are a JSON response of the
same structure.

This lets us skip deserializing the JSON output from PostgreSQL and then
re-serializing it later, which both speeds up responses drastically and
frees up database connections faster, as part of this processing was
done while holding onto the connection.

As we build an array of responses when handling multiple variable sets,
we construct the JSON array by appending bytestrings to a buffer,
prefixed with '[', suffixed with ']', and separated by ','. This can be
error-prone but hopefully won't have to change very often.

CockroachDB, unfortunately, seems to add an 0x01 byte to the start of
the output, which we have to check for and discard. This will never be
valid JSON, so we can discard it safely.

This seems to have fixed a bug in which numbers were deserialized to
floating-point values and then re-serialized, leading to lossiness in
the processing.