Avoid cloning metadata for `Env`. (#41)
### What

The `Env` struct captures the metadata, but it doesn't need to own it as
it's pretty transient.

By capturing a reference instead, we can avoid cloning the metadata,
which is a fairly expensive operation.

On my machine, this has a noticeable impact on benchmark results.

### How

Judicious use of the `&` operator.

Yes, there are lifetimes.