Wrap configuration and state in `Arc`. (#42)
### What

This wraps configuration and state in `Arc`, because they are cloned by
axum per request. [axum encourages this in its
documentation.](https://docs.rs/axum/latest/axum/index.html#sharing-state-with-handlers)

This seems to have a pretty decent impact on benchmark results.

### How

We wrap the raw configuration, configuration, and state in `Arc<...>`,
and fix the code to match. As we treat this data as read-only, that's
really all there is to it.