Remove the SQL AST trace (#45)
### What

We decided that it is not necessary to trace the SQL AST because it is
verbose and the information is already represented in the SQL string in
a more readable format. Since the SQL String is expected to be a direct
translation of the SQL AST losing only very little information (if at
all), this seems acceptable and could potentially make things go a
little bit faster.