Bump ndc-client to v0.1.0-rc.7. (#53)
### What

Time to bump.

This involves bumping ndc-sdk too.

We also need to be able to parse secrets (i.e. connection URIs) which
have the form `{"value": "postgresql://..."}` as well as simply
`"postgresql://..."`.

### How

```
$ sd '^(ndc-sdk = .+, rev = )"[^"]+"(.+)$' '$1"50d29e0"$2' crates/**/Cargo.toml
$ sd '^(ndc-(?:client|test) = .+, tag = )"[^"]+"(.+)$' '$1"v0.1.0-rc.7"$2' crates/**/Cargo.toml
```

I have introduced a wrapper type for a resolved secret, and an
intermediate which handles deserializing from either style.

There was one compilation failure; a function was renamed.