Remove `SingleOrList` in favor of just a single connection uri (#59)
### What

We want to get rid of any untagged enums in our configuration format,
because those are awkward to consume from typescript.

Our only instance of an untagged enum was `SingleOrList`, which was used
to encode that you could have one or more connection uris, and in the
case of a single connection uri you didn't need to wrap it in a list.

Since we currently only support a single database connection anyway I
opted to just accept a single value rather than a list, in order to keep
speculation minimal. If/when we support more than a single connection
string we'll decide what's appropriate, and perhaps bump the
configuration version number accordingly.