Make it easier to run the deployment script without actually deploying. (#58)
### What

I want to make the deployment script easier to extend.

I have:

1. added a `--dry-run` flag which stops it from pushing, which makes it
easier to test, kind of; and
2. changed the parameters so you need to explicitly name the image
rather than computing it.

In order to test the script, I also needed to make it run from this
branch, temporarily. I have therefore added a line to make sure test
failures are only reported to Slack from the `main` branch or a tag.

### How

The usual. Bashing the keyboard to create more bash.