Use the health check when possible in scripts. (#62)
### What

`ndc-postgres check-health` should be more robust than using `nc -z`, as
it makes a call to `/health` and ensures it receives a successful
response, instead of just verifying that the port is open.

### How

I switched out `nc -z localhost $PORT` for `ndc-postgres check-health
--port=$PORT` where possible. There's only one place where it's still
necessary.

In the benchmarks' Docker Compose file, I added health checks so we
could remove the call to `wait-until` in CI. I expect it will fix the
sporadic failures we're seeing.