CI: Ensure that Cargo.lock does not change when building. (#65)
### What

Sadly, `cargo build --locked` isn't enough to check that the lockfile is
up to date, as it will happily run even if there are extra entries in
there.

It does, however, make sure that the file is not updated, which is quite
unhelpful.

### How

Instead of using `--locked`, we build without it, and then use `git` to
figure out if it changed.