In tests, ensure that query responses are valid.
By deserializing the query response to a `QueryResponse` and not just a
`Value`, we can be relatively sure that it conforms to the
specification.

This is especially important now that we pass the JSON directly from the
database to the HTTP response.

This doesn't seem to add any time to the test run, probably because
deserializing to a `Value` is time-consuming too.

The order of some JSON keys in snapshots have changed to conform to the
`QueryResponse` key order.