Add simple constant expressions folding (#75)
### What

When we generate SQL we sometimes generate "messy" code. Specifically,
in `WHERE` expressions we sometimes generate `true AND <expr>` instead
of `<expr>` when combining multiple predicates (where, join, and soon
relationship predicate). I'd like to normalize the expressions in the
AST so that things like `true AND <expr> AND true` are reduced to
`<expr>`, so that test snapshots are easier to read and will not change
with no good reason.

Note that this kind of rewrite does not intend to modify the logic of
the query, nor does it intend to improve performance, since postgres
probably already has a constant folding rewrite of its own. The goal is
to make the generated code nicer, easier to understand, and easier to
test.

Next, I will add handling for the predicate in `PathElement`, and by
making this change I'd expect it to only affect tests that actually
include a meaningful `PathElement`.

### How

A recursive algorithm. Pattern match on three kinds of boolean
expressions - `And`, `Or` and `Not`, apply the rewrite to their
children, then check whether one of the children can be pruned.

This rewrite is applied to the `json_select` generated from the
`translate` function.