Factor out a script to generate a new configuration.
This removes some duplication, but more importantly, it allows us to get
a little more sophisticated.

Now, when a request fails, we print the response status _and_ the body.
Previously, we only printed the status.

We sometimes see a 500 Internal Server Error and don't know why.
Hopefully this will help us get to the bottom of it.