Factor out a script to generate a new configuration. (#87)
### What

This removes some duplication, but more importantly, it allows us to get
a little more sophisticated.

Now, when a request fails, we print the response status _and_ the body.
Previously, we only printed the status.

We sometimes see a 500 Internal Server Error and don't know why.
Hopefully this will help us get to the bottom of it.

### How

Instead of using `curl -f`, which prints a simple error, we capture the
status code using `--write-out '%{http_code}\n'`, and the body using
`--output $OUTPUT_FILE`.

We can then use this to construct a more useful error message.

All other logic stays the same.