use the predicate in PathElement (#81)
### What

We haven't been using the `predicate` in the
[PathElement](https://hasura.github.io/ndc-spec/reference/types.html#pathelement).
This is no good. We need to use it so we don't produce incorrect results
to users.

### How

We generate a boolean Expression for the predicates using the relevant
tables and we `AND` the resulting Expression with the join predicate in
the `WHERE` clause.