Ensure that the deployment configuration conforms to the schema. (#91)
### What

We test that the checked-in `chinook-deployment.json` files conform to
the schema generated by `schemars`. There is a separate test to ensure
these are up-to-date.

### How

The test loads the schema using `jsonschema`, and then validates the
schema. On failure, it iterates through the errors and formats them in a
human-readable way.

I verified this by manually modifying the `chinook-deployment.json`
files to ensure they failed.

Note that this test case is not `async`. This is on purpose.
`jsonschema` will mess with `tokio` runtimes if they exist; it can only
be used in non-async code or with extra effort. As this test does not
need to be async, it was simpler to avoid `tokio`.