Add database information to the execution trace. (#98)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

This adds database information to the execution trace in accordance with
[Semantic Conventions for Database Client Calls][].

We extract the database information from the URL, and query for the
server version number. The version number will be in the libpq format,
e.g. "162000" for v16.2.

### How

I factored out configuration access from creating the pool in order to
re-use the `Url` instead of parsing it twice.

Then I pulled out the database information in the same way as `sqlx`'s
[`PgConnectOptions::parse_from_url`][]. Unfortunately, most of the stuff
here is `pub(crate)` (probably for API stability reasons), so I couldn't
just access the data.

I have not added traces to `/explain`, but I will follow up with those.

[Semantic Conventions for Database Client Calls]:
https://opentelemetry.io/docs/specs/semconv/database/database-spans/
[`PgConnectOptions::parse_from_url`]:
https://github.com/launchbadge/sqlx/blob/49ccc7ca3215fd6e96f38abb528446a894e3a561/sqlx-postgres/src/options/parse.rs#L9-L110