support nulls in variables (#99)
### What

We want to support supplying a null as a value in variables (foreach).

### How

Previously, we didn't support supplying null as a variable (foreach)
because we didn't know the value of types at the time. We have since
shifted to using casts on user values when the type is not a simple
number. We can use the same approach for variables as well and wrap them
in casts. This way we can supply null as a type of Option<String> which
will then be cast to the right type.