Add traces for /explain requests. (#100)
### What

We were missing detailed trace information for requests to `/explain`.
This synchronizes it with requests to `/query`.

### How

I mostly copied the existing traces from the "execute" path. This means
there's nothing in particular keeping them in sync; we will need to do
this manually until we find a better way to do so.