Add `mold` to the Nix flake, and document how to use it. (#106)
### What

[`mold`](https://github.com/rui314/mold) is a drop-in replacement for
the GNU and LLVN linkers which is generally a lot faster. I have tried
it and seen a slight speed boost in recompilation.

This adds it to the Nix flake, and documents how to use it (opt-in).

### How

I added `mold-wrapped` to the Nix flake, which is `mold`, wrapped to
know about all the other Nix stuff. I had to update nixpkgs as this is
pretty recent. People who don't use Nix will need to install it using
their system package manager.

Once you have `mold` installed, you can configure `rustc` to use it. The
least invasive way to do this is to configure it using the `RUSTFLAGS`
environment variable, which should stay local to your machine.

I've added instructions on how to do this in a new _.envrc.example_
file. I kept this file separate from _.envrc_ because I don't want new
examples to trigger direnv to complain about the file changing, and also
so I can make a clear distinction between a code snippet and commentary
without actually running code.