Update ndc-spec to get the build working.
ndc-spec v0.1.0-rc.8 uses workspace inheritance to set the version and
edition, which unfortunately does not work with Crane for Nix (when
using Git targets).

I have inlined the version and edition to solve this, but we also need
to release a new version of ndc-spec and update ndc-sdk.

Until that happens, let's just override it (transitively) using `patch`
in _Cargo.toml_.