Update ndc-spec to get the build working. (#111)
### What

ndc-spec v0.1.0-rc.8 uses workspace inheritance to set the version and
edition, which unfortunately does not work with Crane for Nix (when
using Git targets).

I have inlined the version and edition to solve this, but we also need
to release a new version of ndc-spec and update ndc-sdk.

Until that happens, let's just override it (transitively) using `patch`
in _Cargo.toml_.

### How

[`[patch."..."]`, in
_Cargo.toml_](https://doc.rust-lang.org/cargo/reference/overriding-dependencies.html#the-patch-section),
allows us to hijack a dependency and point it somewhere else. Unlike
just setting the version, this works transitively, so it will also
update ndc-sdk. This is important because it's the dependency build step
which fails in Crane, so we need to ensure we have no reference to the
previous version.

Unfortunately, there's an open bug in Cargo which means that [patch
overrides can't use the same Git
repository](https://github.com/rust-lang/cargo/issues/10756), even if
the commit reference is different. We work around this by adding an
extra "/" to trick Cargo into thinking it's a different repository.