Fix Slack reporting so it always happens, but only for `main`. (#114)
### What

Turns out getting this right is more trouble than you would think.

This should fix the notifications so that they happen on a build failure
on `main`, but no other time.

### How

We need to call `always()` to signify to GitHub Actions that we want to
run even on failure, but we also need to check the branch because
otherwise it'll run on all pushes for all branches.