Stop statically linking binaries when building with Nix. (#126)
### What

It turns out that statically linking with musl increases our response
time under load from (in a very specific benchmark, on my computer)
~10ms to ~30ms per response. Disabling static linking fixes this issue,
bringing ndc-postgres in line with HGE v2.

This feels like a good enough reason not to statically link the code.

We ship the Docker image built with Nix, so this directly impacts
production. It also simplifies development (in theory) because `nix
build` and `cargo build` now have behavior that is more similar.

### How

I deleted the Nix code that turns on static linking.