chore: keep snapshots of old configuration versions (#127)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

As we change our deployment configuration types, we want to ensure we
can still parse old ones. This adds tests to ensure that.

### How

When we run `just generate-chinook-configuration`, we copy the current
deployment into `static/deployment-snapshots`, named by a hash of its
contents.

We add a new test that goes through all files in this folder and tries
to parse them into our `RawConfiguration` type.