Demote E2E failures to Slack messages (#131)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

We want to get the end-to-end tests up and running to ensure consistency
across the repositories, and we want to do it in a minimally intrusive
way (at least for now). Having spoken to @danieljharvey, we decided the
minimally intrusive option was probably just to keep piping build
failures to Slack after the fact, rather than slowing the merge process.

This also removes the as-of-yet-unsolved problem of how you actually
make a breaking change. If E2E failures block commits, how do you
actually change a request/response pair? You can't update the E2E tests
because you haven't made the change yet, and you can't make the change
because the E2E tests fail. Now, you can happily break `main` in one and
then fix the other.

---------

Co-authored-by: Tom Harding <tom.harding@hasura.io>