qualify table names in configuration query (#136)
### What

We want to emit table and views collection names as qualified (with the
schema name) in the configuration generation, so that when there the are
multiple tables with the same name but different schemas, the collection
names won't clash.

### How

In `configuration_query.sql`:

```patch
- rel.relation_name,
+ s.schema_name || '_' || rel.relation_name,
```

The rest of the changes are regenerating the configuration and updating
the tests to use the qualified names.