Remove abstractions over binaries in the build processes. (#135)
### What

Both CI and the Nix flake abstracted over binaries so we could build
more than one of them.

This is now pointless as we only have one. I have removed the
abstractions to make working in these areas easier.

### How

I inlined references to "ndc-postgres" in both the Nix flake and the
GitHub Actions workflows.