add an 'unqualified schemas' list to the configure options (#147)
### What

Currently the configuration server will return table and view names
qualified with their schema, unless the schema is 'public'. In this PR
we make this option configurable - we provide the user with the option
to mention which schemas will not be qualified, with the default being
'public'.

This is an additive change, but the jsonschema changes obviously.

### How

Add a field to `ConfigureOptions` named `unqualified_schemas`, set its
default to be `["public"]`, and pass it as a parameter to the
configuration sql query.

Then fix the configuration tests.