update-multitenant-dep: just tag everyone (#148)
### What

The `update-multitenant-dep` flow seems to fail:
```
GraphQL: Could not resolve to a User with the login of 'github-merge-queue[bot]'. (u000)
```

I think this is because we use merge queues, so when we assign the PR to
`$GITHUB_ACTOR` we actually assign it to `github-merge-queue[bot]`
instead of the developer.

### How

We'll assign it to the team.