Simplify the Nix derivation as much as possible. (#144)
### What

The Nix derivation code we have is quite sophisticated, and does things
we don't need.

I want to clean it up so we can reuse it as much as possible in other
repositories to take advantage of the caching (now we have caching
working well) and the small Docker image size.

### How

I have tried to simplify it in a couple of ways:

1. Extracting out the clever Rust cross-compilation stuff into its own
file which doesn't need to be touched very often.
2. Merging what's left of _cargo-build.nix_ and _ndc-agent.nix_.