Fix Nix cross-compilation by using `callPackage` properly. (#151)
### What

I removed the call to `callPackage` in favor of using `pkgs` directly.
It seems this was a bad idea. I have undone this part of my refactor
and, I hope, made the Nix derivation clearer as a result.

I also added CI jobs for cross-compilation so this doesn't happen again.

### How

Yet more Nix. The key is using `pkgs.callPackage` with the `pkgs` from
_rust.nix_, rather than `import`, to specify the derivation.