Only try to deploy to Staging after building is complete. (#152)
### What

This sets up the deployment workflow to only kick off after the Nix
build finishes. This means that the Docker image will end up in the
cache and won't have to be re-built, which reduces duplicate work.

If this works out, we can re-use this mechanism for other workflows too.

### How

I have replaced the trigger on the `push` event with [GitHub Action's
`workflow_run`
event](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_run),
instructing GitHub Actions to start the deployment workflow after the
"nix build" workflow has completed.

Unfortunately, it won't actually run until this change is merged, as
this is policy for `workflow_run` events.

> **Note:** This event will only trigger a workflow run if the workflow
file is on the default branch.

This means that this is completely untested and may fail
catastrophically. The only way to find out is to merge it.