Optimize foreach queries by constructing a single query (#158)
### What

ndc-spec supports running a query against [variables
sets](https://hasura.github.io/ndc-spec/specification/queries/variables.html).
Meaning that a user can supply a query that contains variables together
with multiple maps of variables. Up until now they way we implemented
this was running the query multiple times (sequentially), once for each
variables set.

In this PR we change an approach and construct a single query that takes
all sets of variables.

We are using this new approach because it appears to be faster.

Before:
![Screenshot from 2023-11-07
17-21-45](https://github.com/hasura/ndc-postgres/assets/8547573/0791bad0-2006-4868-bddf-ec2ab9df1661)
After:
![Screenshot from 2023-11-07
17-21-54](https://github.com/hasura/ndc-postgres/assets/8547573/b7c0700b-0ad8-40c8-8b28-1376b62ee5be)


### How

First, all queries now return arrays instead of a single {rows,
aggregates} object instead of constructing that from rust code.

Second, when the query is a foreach query, we:

1. construct an json object of the form `[{ [<field>: <value>],
"%variable_order": <i> }, ... ]` which we bind to the query
2. Change the query to include select from this object (as a relation)
and `CROSS JOIN LATERAL` the result with the select query which builds
the `{rows, aggregates}` result per query/variables set.
3. Select the relevant variable name as a column from the variables
table when we want to use a variable.
4. Sort the result sets according to the variables sets.
5. Aggregates the results into an array.

Here's an example query. `*` means changed, `+` means new.

```sql
SELECT
  coalesce(json_agg(row_to_json("%1_universe")), '[]') AS "universe" -- * json_agg
FROM
  json_to_recordset(cast($1 as json)) -- +
    AS "%0_variables"("search" varchar, "%variable_order" int) -- +
CROSS JOIN LATERAL -- +
  (
    SELECT
      *
    FROM
      (
        SELECT
          coalesce(json_agg(row_to_json("%2_rows")), '[]') AS "rows"
        FROM
          (
            SELECT
              "%0_Album"."Title" AS "Title"
            FROM
              "public"."Album" AS "%0_Album"
            WHERE
              ("%0_Album"."Title" ~~ cast("%0_variables"."search" as varchar)) -- * instead of just $1
            ORDER BY
              "%0_Album"."AlbumId" ASC
          ) AS "%2_rows"
      ) AS "%2_rows"
    ORDER BY "%0_variables"."%variable_order" -- +
  ) AS "%1_universe"
```

To do this, we have to:

1. Add representation for `CrossJoinLateral` and `JsonToRecordset` in
`sql::ast` and pp them.
2. Change some `sql::helpers` functions to construct the relevant parts.
3. Build the variables table from clause and keep it in the state (since
table names are unique and not magic)
4. Execute a single query with the json object of variables sets
constructed from the `QueryRequest` variables.
5. ???
6. Profit