Inline workspace-level directives to make Crane happy. (#168)
### What

Crane does not properly inline workspace-level directives in all
situations, which means that the version information is not available to
it once it's constructed a derivation for each crate.

This becomes an issue once we use ndc-postgres as a library.

### How

We can work around this problem by not using workspace-level directives,
copying and pasting the information instead.