Optimize order by multiple columns from the same table  (#177)
Big PR. Sorry in advance.

### What

Currently we build a separate join for each element in the order by
element list.
We'd like to optimize this and merge some of the joins in cases where
multiple columns or aggregates from the same table are selected.

We end up implementing order by nested aggregates along the way.

### How

Instead of build sepearate joins for each element in the order by
element list, we first group elements that have the same path together.
Note we separate columns from aggregates because they return different
things.

For example, If I'm querying the `Track` table and I want to sort by:

1. The artist id (found in Album)
2. The track name
3. The album id (found in album)

I create the following grouping:

```hs
[ Columns
  { path: [album]
  , columns: [(0, artist_id), (2, album_id)]
  }
, Columns
  { path: []
  , columns: (1, name)
  }
]
```

Then we go over each group and generate a select query which we will
join, and return the `ColumnReference`s (along with their place in the
order by list and direction) so we can write them in the order by list.

In this code I've had to rewrite quite a few parts to match with the new
types, and I've merged the code paths for aggregates and columns, so we
also get nested aggregate order bys with this work.

So the general flow is:

1. Group columns/aggregates with the same path
2. traverse each group and generate a select (which follows the path and
select the requested columns) as well as list of column references and
what we need to do with them (aggregate them, how to place them in the
select list)
3. Wrap each path traversal with a select that selects the requested
columns and applies the aggregations if needed
4. Place the result `ColumnReference`s in the order by list

Note that for empty paths we just return the columns, and fail if
aggregates are requested on empty paths.