faster mutation tests (#190)
### What

Mutation tests are slow on my machine because populating the fresh
database takes a long time.
Instead of creating a new database and then running the chinook script,
we want to copy the chinook database, which is faster.

### How

We prepare a copy of the postgres database after running the chinook
script so that we can use later to copy the template to a fresh database
we can run mutations on.