Native query mutations (#189)
### What

We want to support the
[spec](https://hasura.github.io/ndc-spec/specification/mutations/index.html)
definition for mutations over procedures by allowing users to define
native queries that can do mutations.

Native Query Mutations should return the columns of the table using a
`RETURNING` clause,
so we can select those when building the response query.

![Screenshot from 2023-11-28
17-24-58](https://github.com/hasura/ndc-postgres/assets/8547573/9451327a-9ab2-488e-82ae-65a23f3f63b2)


### How

tl;dr, we:

1. revise the `ExecutionPlan` to support a list of mutations as well as
a query.
2. translate each mutation to something like `with "nq" as (<native
query>) select <fields>, <affected row> from <selects>`.
3. implement the execution of the mutations list by running them on
after the other in a transaction and concatenate the results.
4. implement the `/mutation` endpoint to call the relevant parts of the
translation and execution.

---------

Co-authored-by: Philip Lykke Carlsen <plcplc@gmail.com>