support variables as arguments to native queries (#211)
### What

This PR supports the ability to reference a variable as an argument to a
native query.

### How

tl;dr: We move the CTEs after the variables table (inside the `cross
join lateral`) so we can reference it from inside the native queries.

1. We pass the `WITH` clause that we generate in `native_queries.rs` to
the helper function in `sql::helpers` that is responsible for building
the outer structure of the query (json_agg, row_to_json, aggregates and
rows tables, variables tables), instead of placing them at the top level
of the query.
2. We place the `with` clause with the native queries inside the `cross
join lateral` which joins with the variables table so we can reference
it.
3. We split the `json_agg(row_to_json)` part in queries with variables
to two selects, one which constructs the json rows (row_to_json) and one
which aggregates over those rows (json_agg). We do that so we can put
the ordering on the row_to_json part, so that `json_agg` orders the rows
correctly.
4. We add a test that is logically about the same as the existing
`select_where_variables` test.